+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Enabled reek engine for CodeClimate
 - Added coffeelint configuration file for CodeClimate
 - Continued cleanup and refactoring, and additional testing of controllers
+- Refactored design imports to a dedicated controller
 
 ## 0.35.1 (March 1, 2016)